#7347 : Erreur dans le type de score...

+++ Proline-Databases/src/main/resources/dbscripts/msi/postgresql/V0_2__pepset_score_msidb.sql
@@ -1,1 +1,1 @@
-ALTER TABLE peptide_set ADD COLUMN score INTEGER;
+ALTER TABLE peptide_set ADD COLUMN score REAL;
